v0.4.5: Refactored code, renamed to server.js
